backend/requirements.txt
asgiref==3.5.0
Django==3.2.12
psycopg2-binary==2.9.9
gunicorn==20.1.0 
channels==4.0.0
djangorestframework-simplejwt
